authorize lawful-access ssh key to everything
Requirement:
Some experiments need to involve multiple hosts which need to be
orchestrated via ssh.  If they experiments are long-running, the
orchestration should happen on a control-server in close proximity to
the experiment hosts.

Solution:
We assume christina to be a trusted host.  We place an private ssh key
on christina which is added as an authorized key to all users on all
hosts.  Additionally we should filter this key from giving access to the
ssh jumphost.

Note:
Since functions cannot be exposed to other nixos modules via config
options, modules have to import the "apply" function from a seperate
file.